+++ src/components/PageA.tsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Axios from 'axios';
 import { DataType } from '../types/DataType';
 import '../App.css';
+import Header from "./Header"
 
 const App = () => {
     const [data, setData] = useState<DataType>([]);
@@ -15,6 +16,7 @@ const App = () => {
 
     return (
         <div>
+            <Header />
             <h1>PageA</h1>
             <table>
                 <thead>
